Revert "rename component files to ease importing"
This reverts commit c8d79cd51d24526bd0bd76d47a63ae369987c433.
