+++ src/components/burger-ingredients/tab-bar/action-tab/action-tab.jsx
